@@ -1,149 +1,153 @@
 <?xml version="1.0" encoding="UTF-8"?>
-
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0
+                             https://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
 
     <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.6.1</version>
+        <version>3.2.4</version>
         <relativePath/>
     </parent>
-
-    <modelVersion>4.0.0</modelVersion>
 
     <groupId>com.easypan</groupId>
     <artifactId>easypan</artifactId>
     <version>1.0</version>
     <packaging>jar</packaging>
-    <name>easypan</name>
-    <description>easypan</description>
 
     <properties>
-        <java.version>1.8</java.version>
+        <java.version>17</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <maven.compiler.source>1.8</maven.compiler.source>
-        <maven.compiler.target>1.8</maven.compiler.target>
         <skipTests>true</skipTests>
 
-        <springboot.version>2.6.1</springboot.version>
-        <redisson.version>3.25.2</redisson.version>
-        <mybatis.version>1.3.2</mybatis.version>
-        <logback.version>1.2.10</logback.version>
-        <mysql.version>8.0.23</mysql.version>
-        <aspectjweaver.version>1.9.4</aspectjweaver.version>
-        <okhttp3.version>3.2.0</okhttp3.version>
-        <fastjson.version>1.2.66</fastjson.version>
-        <commons.lang3.version>3.4</commons.lang3.version>
-        <commons.codec.version>1.9</commons.codec.version>
-        <commons.io.version>2.5</commons.io.version>
+        <!-- 仅对 MyBatis Starter 版本单独管理 -->
+        <mybatis.spring.boot.version>3.0.4</mybatis.spring.boot.version>
 
+        <!-- 保留原有的第三方版本属性 -->
+        <redisson.version>3.25.2</redisson.version>
+        <mysql.version>8.0.33</mysql.version>
+        <aspectjweaver.version>1.9.21</aspectjweaver.version>
+        <okhttp3.version>4.12.0</okhttp3.version>
+        <fastjson.version>2.0.47</fastjson.version>
+        <commons.lang3.version>3.14.0</commons.lang3.version>
+        <commons.codec.version>1.16.0</commons.codec.version>
+        <commons.io.version>2.15.1</commons.io.version>
+        <freemarker.version>2.3.32</freemarker.version>
     </properties>
 
     <dependencies>
+        <!-- Spring Web -->
         <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
             <exclusions>
                 <exclusion>
                     <groupId>ch.qos.logback</groupId>
-                    <artifactId>logback-classic</artifactId>
-                </exclusion>
-                <exclusion>
-                    <groupId>ch.qos.logback</groupId>
                     <artifactId>logback-core</artifactId>
                 </exclusion>
             </exclusions>
         </dependency>
-	
-	    <!--邮件发送-->
-	    <dependency>
+
+        <!-- Actuator（可选，加上后可启用调度 Metrics） -->
+        <dependency>
             <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-mail</artifactId>
-            <version>${springboot.version}</version>
-        </dependency>
-	    <!--redis -->
-	    <dependency>
-            <groupId>org.springframework.boot</groupId>
-            <artifactId>spring-boot-starter-data-redis</artifactId>
-            <version>${springboot.version}</version>
+            <artifactId>spring-boot-starter-actuator</artifactId>
         </dependency>
 
-        <!-- redisson-->
+        <!-- 邮件、Redis、Redisson -->
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-mail</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.springframework.boot</groupId>
+            <artifactId>spring-boot-starter-data-redis</artifactId>
+        </dependency>
         <dependency>
             <groupId>org.redisson</groupId>
             <artifactId>redisson-spring-boot-starter</artifactId>
-            <version>${redisson.version}</version> <!-- 或最新版本 -->
+            <version>${redisson.version}</version>
         </dependency>
 
+        <!-- Lombok -->
+        <dependency>
+            <groupId>org.projectlombok</groupId>
+            <artifactId>lombok</artifactId>
+            <scope>provided</scope>
+        </dependency>
 
-        <!--mybatis-->
-	    <dependency>
+        <!-- —— 正确引入 MyBatis Spring Boot Starter —— -->
+        <dependency>
             <groupId>org.mybatis.spring.boot</groupId>
             <artifactId>mybatis-spring-boot-starter</artifactId>
-            <version>${mybatis.version}</version>
+            <version>${mybatis.spring.boot.version}</version>
         </dependency>
-	
-	
-	    <!-- 数据库-->
-	    <dependency>
+
+        <!-- 数据库驱动 -->
+        <dependency>
             <groupId>mysql</groupId>
             <artifactId>mysql-connector-java</artifactId>
             <version>${mysql.version}</version>
         </dependency>
-	
-	
-	    <!-- 日志版本 -->
-	    <dependency>
+
+        <!-- Logging -->
+        <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
-            <version>${logback.version}</version>
+            <version>1.4.14</version>
         </dependency>
         <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-core</artifactId>
-            <version>${logback.version}</version>
+            <version>1.4.14</version>
         </dependency>
-	
-	    <!--切面-->
-	    <dependency>
+
+        <!-- AOP -->
+        <dependency>
             <groupId>org.aspectj</groupId>
             <artifactId>aspectjweaver</artifactId>
             <version>${aspectjweaver.version}</version>
         </dependency>
-	
-	    <!--okhttp-->
-	    <dependency>
+
+        <!-- HTTP 客户端、JSON、Commons、模板 -->
+        <dependency>
             <groupId>com.squareup.okhttp3</groupId>
             <artifactId>okhttp</artifactId>
             <version>${okhttp3.version}</version>
         </dependency>
-	
-	    <!--fastjson-->
-	    <dependency>
+        <!-- FastJSON -->
+        <dependency>
             <groupId>com.alibaba</groupId>
             <artifactId>fastjson</artifactId>
             <version>${fastjson.version}</version>
         </dependency>
-	
-	    <!--apache common-->
-	    <dependency>
+        <!-- Apache Commons -->
+        <dependency>
             <groupId>org.apache.commons</groupId>
             <artifactId>commons-lang3</artifactId>
             <version>${commons.lang3.version}</version>
         </dependency>
-
+        <!-- Apache Commons Codec -->
         <dependency>
             <groupId>commons-codec</groupId>
             <artifactId>commons-codec</artifactId>
             <version>${commons.codec.version}</version>
         </dependency>
-
+        <!-- Apache Commons IO -->
         <dependency>
             <groupId>commons-io</groupId>
             <artifactId>commons-io</artifactId>
             <version>${commons.io.version}</version>
         </dependency>
+        <!-- Freemarker -->
+        <dependency>
+            <groupId>org.freemarker</groupId>
+            <artifactId>freemarker</artifactId>
+            <version>${freemarker.version}</version>
+        </dependency>
+
     </dependencies>
 
     <build>
