@@ -3,6 +3,7 @@ package com.easypan;
 import com.easypan.entity.config.AppConfig;
 import com.easypan.entity.constants.Constants;
 import com.easypan.spring.ApplicationContextProvider;
+import jakarta.servlet.MultipartConfigElement;
 import org.mybatis.spring.annotation.MapperScan;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
@@ -13,12 +14,12 @@ import org.springframework.scheduling.annotation.EnableAsync;
 import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.transaction.annotation.EnableTransactionManagement;
 
-import javax.servlet.MultipartConfigElement;
+
 
 
 @EnableAsync
 @SpringBootApplication(scanBasePackages = {"com.easypan"})
-@MapperScan(basePackages = {"com.easypan.mappers"})
+@MapperScan("com.easypan.mappers")
 @EnableTransactionManagement
 @EnableScheduling
 public class EasyPanApplication {
