@@ -1,9 +1,9 @@
 package com.easypan.controller;
+
 import com.easypan.entity.enums.ResponseCodeEnum;
 import com.easypan.entity.vo.ResponseVO;
 import com.easypan.exception.BusinessException;
-
-
+import jakarta.servlet.http.HttpServletRequest;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.dao.DuplicateKeyException;
@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import org.springframework.web.method.annotation.MethodArgumentTypeMismatchException;
 import org.springframework.web.servlet.NoHandlerFoundException;
 
-import javax.servlet.http.HttpServletRequest;
+
 
 @RestControllerAdvice
 public class AGlobalExceptionHandlerController extends ABaseController {
